refactor: ChannelCard 컴포넌트를 Card 컴포넌트로 대체
ChannelCard 컴포넌트를 Card 컴포넌트로 대체하여 ChannelDetail 컴포넌트
만을 사용하도록  변경
