--- frontend/src/components/Channel/ChannelCard/ChannelCard.stories.jsx
@@ -1,8 +1,0 @@
-import ChannelCard from './ChannelCard';
-
-export default {
-    title: 'ChannelCard',
-    component: ChannelCard,
-};
-
-export const Default = ChannelCard.bind();
