+++ frontend/src/components/Channel/ChannelDetail/Avatar.jsx
