+++ frontend/src/components/Channel/ChannelDetail/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './ChannelDetail';
