+++ frontend/src/components/Channel/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './Channel';
